fixed 4 comments
