+++ store/action/cart.js
@@ -1,4 +1,5 @@
 export const ADD_PRODUCT = 'ADD_PRODUCT';
+
 export const addProduct = (product) => {
   return { type: ADD_PRODUCT, product: product };
 };
